+++ packages/curvenote-cli/src/uploads/utils.ts
@@ -1,23 +1,13 @@
 import type { ISession } from '../session/types.js';
 import { createHash } from 'node:crypto';
-import cliProgress from 'cli-progress';
 import fs from 'node:fs';
 import mime from 'mime-types';
 import { tic } from 'myst-cli-utils';
 import path from 'node:path';
-import pLimit from 'p-limit';
-import type { SiteDeployRequest, SiteUploadRequest, SiteUploadResponse } from '@curvenote/blocks';
-import { addOxaTransformersToOpts } from '../utils/index.js';
 import type { FileInfo, FileUpload, FromTo } from './types.js';
-
-export const siteCommandWrapper =
-  (
-    siteCommand: (session: ISession, opts: Record<string, any>) => Promise<any>,
-    defaultOptions: Record<string, any>,
-  ) =>
-  async (session: ISession, opts: Record<string, any>) => {
-    await siteCommand(session, addOxaTransformersToOpts(session, { ...defaultOptions, ...opts }));
-  };
+import cliProgress from 'cli-progress';
+import pLimit from 'p-limit';
+import type { SiteUploadRequest, SiteUploadResponse } from '@curvenote/blocks';
 
 export function listFolderContents(session: ISession, from: string, to = ''): FromTo[] {
   const directory = fs.readdirSync(from);
@@ -40,7 +30,7 @@ export function listFolderContents(session: ISession, from: string, to = ''): Fr
   return [...outputFiles, ...outputFolders];
 }
 
-export async function prepareFileForUpload(from: string, to: string): Promise<FileInfo> {
+export function makeFileInfo(from: string, to: string): FileInfo {
   const content = fs.readFileSync(from).toString();
   const stats = fs.statSync(from);
   const md5 = createHash('md5').update(content).digest('hex');
@@ -83,9 +73,7 @@ export async function prepareUploadRequest(session: ISession) {
   const filesToUpload = listFolderContents(session, session.sitePath());
   session.log.info(`🔬 Preparing to upload ${filesToUpload.length} files`);
 
-  const files = await Promise.all(
-    filesToUpload.map(({ from, to }) => prepareFileForUpload(from, to)),
-  );
+  const files = filesToUpload.map(({ from, to }) => makeFileInfo(from, to));
 
   const uploadRequest: SiteUploadRequest = {
     files: files.map(({ md5, size, contentType, to }) => ({
@@ -140,83 +128,4 @@ export async function performFileUploads(
 
   const cdnKey = uploadTargets.id;
   return { cdnKey };
-}
-
-/**
- * Upload content to the (private) staging bucket
- *
- * @param session
- * @param opts
- * @returns cdnkey and filepaths for deployment
- */
-export async function uploadContent(session: ISession, opts?: { ci?: boolean }) {
-  const { files, uploadRequest } = await prepareUploadRequest(session);
-  const { json: uploadTargets } = await session.post<SiteUploadResponse>('/sites/upload', {
-    ...uploadRequest,
-  });
-  const { cdnKey } = await performFileUploads(session, files, uploadTargets, opts);
-  return { cdnKey, filepaths: files.map(({ to }) => ({ path: to })) };
-}
-
-/**
- * Deploy content to the public/private CDN based on the usePublicCdn argument
- *
- * @param session
- * @param usePublicCdn whether to deploy to the public CDN or default private CDN
- * @param cdnKey
- * @param filepaths
- * @returns cdnkey
- */
-export async function deployContent(
-  session: ISession,
-  privacy: { public: boolean },
-  cdnKey: string,
-  filepaths: { path: string }[],
-) {
-  const toc = tic();
-  const deployRequest: SiteDeployRequest = {
-    public: privacy.public,
-    id: cdnKey,
-    files: filepaths,
-  };
-  const deployResp = await session.post('/sites/deploy', deployRequest);
-
-  if (deployResp.ok) {
-    session.log.info(toc(`🚀 Deployed ${filepaths.length} files in %s.`));
-    session.log.debug(`CDN key: ${cdnKey}`);
-  } else {
-    throw new Error('Deployment failed: Please contact support@curvenote.com!');
-  }
-  return cdnKey;
-}
-
-/**
- * Perform PRIVATE CDN upload and deployment
- *
- * @param session
- * @param opts
- * @returns cdnkey
- */
-export async function uploadContentAndDeployToPrivateCdn(
-  session: ISession,
-  opts?: { ci?: boolean },
-) {
-  const { cdnKey, filepaths } = await uploadContent(session, opts);
-  // TODO bespoke private CDN
-  return await deployContent(session, { public: false }, cdnKey, filepaths);
-}
-
-/**
- * Perform PUBLIC CDN upload and deployment
- *
- * @param session
- * @param opts
- * @returns cdnkey
- */
-export async function uploadContentAndDeployToPublicCdn(
-  session: ISession,
-  opts?: { ci?: boolean },
-) {
-  const { cdnKey, filepaths } = await uploadContent(session, opts);
-  return await deployContent(session, { public: true }, cdnKey, filepaths);
 }
