+++ packages/curvenote-cli/src/utils/types.ts
@@ -1,21 +1,3 @@
-export type FromTo = {
-  from: string;
-  to: string;
-};
-
-export type FileInfo = {
-  from: string;
-  to: string;
-  md5: string;
-  size: number;
-  contentType: string;
-};
-
-export type FileUpload = FileInfo & {
-  bucket: string;
-  signedUrl: string;
-};
-
 export interface WorkBody {
   key: string;
   cdn: string;
